+++ packages/eas-cli/src/project/__tests__/projectUtils-test.ts
@@ -1,18 +1,13 @@
 import { getConfig, modifyConfigAsync } from '@expo/config';
 import { vol } from 'memfs';
 
-import { confirmAsync } from '../../prompts';
 import { Actor, getUserAsync } from '../../user/User';
-import {
-  ensureProjectExistsAsync,
-  findProjectIdByAccountNameAndSlugNullableAsync,
-} from '../ensureProjectExists';
+import { ensureProjectExistsAsync } from '../ensureProjectExists';
 import {
   findProjectRootAsync,
   getProjectAccountName,
   getProjectAccountNameAsync,
   getProjectIdAsync,
-  promptToCreateProjectIfNotExistsAsync,
 } from '../projectUtils';
 
 jest.mock('@expo/config');
@@ -21,6 +16,10 @@ jest.mock('fs');
 jest.mock('../../prompts');
 jest.mock('../../user/User');
 jest.mock('../ensureProjectExists');
+
+beforeEach(() => {
+  jest.resetAllMocks();
+});
 
 describe(findProjectRootAsync, () => {
   beforeEach(() => {
@@ -35,7 +34,7 @@ describe(findProjectRootAsync, () => {
       '/app'
     );
     await expect(findProjectRootAsync({ cwd: '/app' })).rejects.toThrow(
-      'Please run this command inside a project directory.'
+      'Run this command inside a project directory.'
     );
   });
 
@@ -182,76 +181,22 @@ describe(getProjectAccountNameAsync, () => {
   });
 });
 
-describe(promptToCreateProjectIfNotExistsAsync, () => {
-  beforeEach(() => {
-    jest.mocked(getUserAsync).mockReset();
-    jest.mocked(findProjectIdByAccountNameAndSlugNullableAsync).mockReset();
-  });
-
-  it('returns the existing project', async () => {
-    jest.mocked(getUserAsync).mockImplementation(
-      async (): Promise<Actor> => ({
-        __typename: 'User',
-        id: 'user_id',
-        username: 'quin',
-        accounts: [{ id: 'account_id_1', name: 'quin' }],
-        isExpoAdmin: false,
-      })
-    );
-    jest
-      .mocked(findProjectIdByAccountNameAndSlugNullableAsync)
-      .mockImplementation(async () => 'already-existing-project-id');
-    const exp: any = {};
-    const projectId = await promptToCreateProjectIfNotExistsAsync(exp);
-    expect(projectId).toBe('already-existing-project-id');
-  });
-  it('makes a new project if none exists', async () => {
-    jest.mock('../../prompts');
-    jest.mocked(getUserAsync).mockImplementation(
-      async (): Promise<Actor> => ({
-        __typename: 'User',
-        id: 'user_id',
-        username: 'quin',
-        accounts: [{ id: 'account_id_1', name: 'quin' }],
-        isExpoAdmin: false,
-      })
-    );
-    jest
-      .mocked(findProjectIdByAccountNameAndSlugNullableAsync)
-      .mockImplementation(async () => null);
-    jest.mocked(ensureProjectExistsAsync).mockImplementation(async () => 'new-project-id');
-
-    // user agrees to make new project
-    jest.mocked(confirmAsync).mockImplementation(async () => true);
-    const exp: any = {};
-    const projectId = await promptToCreateProjectIfNotExistsAsync(exp);
-    expect(projectId).toBe('new-project-id');
-  });
-  it('aborts making a new project if the user declines', async () => {
-    jest.mock('../../prompts');
-    jest.mocked(getUserAsync).mockImplementation(
-      async (): Promise<Actor> => ({
-        __typename: 'User',
-        id: 'user_id',
-        username: 'quin',
-        accounts: [{ id: 'account_id_1', name: 'quin' }],
-        isExpoAdmin: false,
-      })
-    );
-    jest
-      .mocked(findProjectIdByAccountNameAndSlugNullableAsync)
-      .mockImplementation(async () => null);
-    jest.mocked(ensureProjectExistsAsync).mockImplementation(async () => 'new-project-id');
-
-    // user doesnt agree to make new project
-    jest.mocked(confirmAsync).mockImplementation(async () => false);
-    const exp: any = {};
-    const projectId = await promptToCreateProjectIfNotExistsAsync(exp);
-    expect(projectId).toBe(null);
-  });
-});
-
 describe(getProjectIdAsync, () => {
+  beforeEach(() => {
+    jest.mocked(getUserAsync).mockImplementation(
+      async (): Promise<Actor> => ({
+        __typename: 'User',
+        id: 'user_id',
+        username: 'notnotbrent',
+        accounts: [
+          { id: 'account_id_1', name: 'notnotbrent' },
+          { id: 'account_id_2', name: 'dominik' },
+        ],
+        isExpoAdmin: false,
+      })
+    );
+  });
+
   it('gets the project ID from app config if exists', async () => {
     await expect(
       getProjectIdAsync({ name: 'test', slug: 'test', extra: { eas: { projectId: '1234' } } })
@@ -273,5 +218,17 @@ describe(getProjectIdAsync, () => {
     expect(modifyConfigAsync).toHaveBeenCalledWith('/app', {
       extra: { eas: { projectId: '2345' } },
     });
+  });
+
+  it('does not throw if writing the ID back to the config fails', async () => {
+    jest.mocked(getConfig).mockReturnValue({ exp: { name: 'test', slug: 'test' } } as any);
+    jest.mocked(modifyConfigAsync).mockResolvedValue({
+      type: 'fail',
+      config: null,
+    });
+    jest.mocked(ensureProjectExistsAsync).mockImplementation(async () => '4567');
+
+    const projectId = await getProjectIdAsync({ name: 'test', slug: 'test' }, {}, { cwd: '/app' });
+    expect(projectId).toBe('4567');
   });
 });